+++ src/zpl.rs
@@ -17,17 +17,17 @@ pub const ICMP_INTERACION_REQUEST_RESPONSE: &str = "request-response";
 pub const ICMP_INTERACTION_ONESHOT: &str = "oneshot";
 
 pub const VISA_SERVICE_CN: &str = "vs.zpr";
-pub const ZPR_ADDR_ATTR: &str = "zpr.addr";
 
 pub const DEFAULT_TS_PREFIX: &str = "device.zpr.adapter";
+
 pub const DEFAULT_ATTR: &str = "cn";
-pub const ADAPTER_CN_ATTR: &str = KATTR_CN;
 
 // TODO: What is up with this odd name? Why not '/zpr/visaservice'?
 pub const VS_SERVICE_NAME: &str = "/zpr/$$zpr/visaservice";
 
 pub const KATTR_ROLE: &str = "zpr.role";
 pub const KATTR_CN: &str = "device.zpr.adapter.cn";
+pub const KATTR_ADDR: &str = "zpr.addr";
 
 pub const ATTR_DOMAIN_SERVICE: &str = "service";
 pub const ATTR_DOMAIN_USER: &str = "user";